move HoneycombInstrumentedComposable to new package
