+++ example/src/main/java/io/honeycomb/opentelemetry/android/example/MainActivity.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import io.honeycomb.opentelemetry.android.LocalOpenTelemetryRum
+import io.honeycomb.opentelemetry.android.compose.LocalOpenTelemetryRum
 import io.honeycomb.opentelemetry.android.example.ui.theme.HoneycombOpenTelemetryAndroidTheme
 import io.opentelemetry.android.OpenTelemetryRum
